requirements.txt
amqp==2.4.2
backcall==0.1.0
billiard==3.5.0.5
celery==4.2.2
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
Django==2.1.7
djangorestframework==3.9.2
idna==2.8
ipdb==0.12
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
kombu==4.3.0
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.9
redis==3.2.1
requests==2.21.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
vine==1.3.0
wcwidth==0.1.7
